Manage users view

+++ run.sh
@@ -5,6 +5,11 @@ RESET_ENVIRONMENT=$1
 RESET_DB=$2
 
 
+echo -e "\033[1;32mStopping running containers...\033[0m"
+docker compose down
+clear # Clear the terminal for better readability
+
+
 function reset_db {
     echo -e "\033[1;32mResetting database...\033[0m"
     COMPOSE_PROFILES="mongo-nass" docker compose down -v
@@ -21,6 +26,14 @@ function reset_db {
 
 
 if [ "$RESET_ENVIRONMENT" = "true" ]; then
+
+    # Ask for confirmation
+    read -p "Are you sure you want to reset the entire environment? This will delete all Docker containers and volumes, including mongo-nass (main) database. (yes/no): " CONFIRM
+    if [ "$CONFIRM" != "yes" ]; then
+        echo -e "\033[1;31mEnvironment reset cancelled.\033[0m"
+        exit 1
+    fi
+
     echo -e "\033[1;32mCleaning up existing Docker containers and volumes...\033[0m"
     rm -rf ./backend/auth-data
     
@@ -30,13 +43,13 @@ if [ "$RESET_ENVIRONMENT" = "true" ]; then
     fi
     
     docker system prune -a -f --volumes
-    COMPOSE_PROFILES="mongo-nass,auth-api,mongo-express,test-services,test-global,dummy-api" docker compose down -v
+    COMPOSE_PROFILES="mongo-nass,auth-api,mongo-express,test-services,test-global,dummy-api, mongo-nass-test,auth-api-test" docker compose down -v
     
     echo -e "\033[1;32mStarting MongoDB...\033[0m"
     COMPOSE_PROFILES="mongo-nass" docker compose up -d mongo-nass
 else
     echo -e "\033[1;32mStopping and removing existing containers, volumes, and networks...\033[0m"
-    COMPOSE_PROFILES="mongo-nass,auth-api,mongo-express,test-services,test-global,dummy-api" docker compose down
+    COMPOSE_PROFILES="mongo-nass,auth-api,mongo-express,test-services,test-global,dummy-api, mongo-nass-test,auth-api-test" docker compose down
 fi
 
 if [ "$RESET_DB" = "true" ]; then
@@ -61,3 +74,4 @@ echo -e "\033[1;32mAll services should be running. Showing logs...\033[0m"
 
 # Run all logs with profiles
 COMPOSE_PROFILES="mongo-nass,auth-api,mongo-express,dummy-api" docker compose logs -f
+
